Extract BlockTransactionChecker class
